feat: add functionality in accordance with the technical task of sprint number 12

+++ src/main/java/ru/yandex/practicum/filmorate/controller/GenreController.java
@@ -23,7 +23,7 @@ public class GenreController {
     }
 
     @GetMapping("/{id}")
-    public Genre getGenre(@PathVariable Long id) {
+    public Genre getGenre(@PathVariable Integer id) {
         return genreService.getGenre(id);
     }
 }
